% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete.R
\name{get_athlete}
\alias{get_athlete}
\title{Get data for a specific athlete}
\usage{
get_athlete(fn, sn, club = NULL)
}
\arguments{
\item{fn}{character. Athletes forename.}

\item{sn}{character. Athletes surname.}

\item{club}{character. The registered club for the athlete}
}
\description{
\code{get_athlete} returns a dataframe with a specified athletes performances. May return more than one athlete
if a club is not specified and there are multiple athletes with the same name.
}
