% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_club_athletes.R
\name{get_club_athletes}
\alias{get_club_athletes}
\title{Find athletes in a club}
\usage{
get_club_athletes(club)
}
\arguments{
\item{club}{string. Club to search}
}
\description{
\code{get_club_athletes} returns all athletes in a specified club.
}
\examples{
\dontrun{
get_club_athletes(club = "Cardiff")
}

}
