% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Get performances for a given year/event/gender}
\usage{
get_event(event, agegroup = "ALL", gender = "M", year, top_n = NULL)
}
\arguments{
\item{event}{character string. The event you want to return.}

\item{agegroup}{character string. Specify age group e.g., "U20", defaults to "ALL".}

\item{year}{character. The year to return in YYYY format, e.g., "2020".}

\item{top_n}{integer. Return top n rows.}

\item{sex}{character. Either "M" for males or "F" for females.}
}
\description{
Get performances for a given year/event/gender
}
